backend ready, i can build frontend now

+++ src/main/java/com/leoric/ecommerceshopbe/services/impl/DealServiceImpl.java
@@ -1,8 +1,11 @@
 package com.leoric.ecommerceshopbe.services.impl;
 
 import com.leoric.ecommerceshopbe.models.Deal;
+import com.leoric.ecommerceshopbe.models.HomeCategory;
 import com.leoric.ecommerceshopbe.repositories.DealRepository;
+import com.leoric.ecommerceshopbe.repositories.HomeCategoryRepository;
 import com.leoric.ecommerceshopbe.services.interfaces.DealService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,30 +16,45 @@ import java.util.List;
 public class DealServiceImpl implements DealService {
 
     private final DealRepository dealRepository;
+    private final HomeCategoryRepository homeCategoryRepository;
 
     @Override
-    public Deal createDeal(Deal deal) {
-        return null;
-    }
-
-    @Override
-    public List<Deal> findAll() {
+    public List<Deal> getAllDeals() {
         return dealRepository.findAll();
     }
 
     @Override
-    public Deal findById(Long id) {
-        return dealRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Deal not found"));
+    public Deal createDeal(Deal deal) {
+        HomeCategory category = homeCategoryRepository.findById(deal.getCategory().getId()).orElse(null);
+        Deal newDeal = new Deal();
+        newDeal.setCategory(category);
+        newDeal.setDiscount(deal.getDiscount());
+        return dealRepository.save(newDeal);
     }
 
     @Override
-    public Deal save(Deal entity) {
-        return dealRepository.save(entity);
+    public Deal updateDealById(Long id, Deal deal) {
+        Deal oldDeal = dealRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Deal not found"));
+        HomeCategory category = homeCategoryRepository.findById(deal.getCategory().getId()).orElse(null);
+        if (deal.getDiscount() != null) {
+            oldDeal.setDiscount(deal.getDiscount());
+        }
+        if (category != null) {
+            oldDeal.setCategory(category);
+        }
+        return dealRepository.save(oldDeal);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteDealById(Long id) {
+        if (!dealRepository.existsById(id)) {
+            throw new EntityNotFoundException("Deal not found");
+        }
         dealRepository.deleteById(id);
+    }
+
+    @Override
+    public Deal getDealById(Long id) {
+        return dealRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Deal not found"));
     }
 }
